requirements.txt
amqp                         5.3.1
annotated-types              0.7.0
asgiref                      3.8.1
async-timeout                5.0.1
attrs                        24.2.0
billiard                     4.2.1
cachetools                   5.5.0
celery                       5.4.0
certifi                      2024.8.30
cffi                         1.17.1
charset-normalizer           3.4.0
click                        8.1.7
click-didyoumean             0.3.1
click-plugins                1.1.1
click-repl                   0.3.0
cryptography                 44.0.0
Django                       5.1.4
google-ai-generativelanguage 0.6.10
google-api-core              2.23.0
google-api-python-client     2.154.0
google-auth                  2.36.0
google-auth-httplib2         0.2.0
google-generativeai          0.8.3
googleapis-common-protos     1.66.0
grpcio                       1.68.1
grpcio-status                1.68.1
h11                          0.14.0
httplib2                     0.22.0
idna                         3.10
kombu                        5.4.2
outcome                      1.3.0.post0
packaging                    24.2
pip                          22.3
prompt_toolkit               3.0.48
proto-plus                   1.25.0
protobuf                     5.29.1
pyasn1                       0.6.1
pyasn1_modules               0.4.1
PyAudio                      0.2.14
pycparser                    2.22
pydantic                     2.10.3
pydantic_core                2.27.1
pyOpenSSL                    24.3.0
pyparsing                    3.2.0
python-dateutil              2.9.0.post0
python-dotenv                1.0.1
redis                        5.2.1
requests                     2.32.3
rsa                          4.9
selenium                     4.0.0
setuptools                   65.5.0
six                          1.17.0
sniffio                      1.3.1
sortedcontainers             2.4.0
SpeechRecognition            3.8.1
sqlparse                     0.5.2
tqdm                         4.67.1
trio                         0.27.0
trio-websocket               0.11.1
typing_extensions            4.12.2
tzdata                       2024.2
uritemplate                  4.1.1
urllib3                      1.26.20
urllib3-secure-extra         0.1.0
vine                         5.1.0
wcwidth                      0.2.13
webdriver-manager            4.0.2
wsproto                      1.2.0